requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
oauthlib==3.2.1
python-dotenv==0.21.0
python-env==1.0.0
requests==2.28.1
requests-oauthlib==1.3.1
urllib3==1.26.12
